Lab-4-Model Parameters & Output Parsing Lab/requirements.txt
# Pydantic for structured output parsing
pydantic==2.12.5

# Environment and Jupyter
python-dotenv

# Token cost tracking (100+ models, live pricing from api.litellm.ai)
litellm==1.82.5
